refactor: index to expose in utils
