+++ src/utils/expose.ts
@@ -1,8 +1,6 @@
 export * as Bytes from './bytes'
 export * as Data from './data'
-export * as Error from './error'
 export * as File from './file'
-export * as Headers from './headers'
 export * as Hex from './hex'
-export * as SafeAxios from './safeAxios'
 export * as Tar from './tar'
+export * from './error'
